refactor: reorganize frontend components
