+++ ourportfolios/auth_config.py
@@ -5,13 +5,14 @@ auth_state.py imports from this file; pages import from auth_state — no cycle.
 """
 
 import os
+
 from dotenv import load_dotenv
 
 load_dotenv()
 
 SUPABASE_URL: str = os.environ.get("SUPABASE_URL", "")
 SUPABASE_ANON_KEY: str = os.environ.get("SUPABASE_ANON_KEY", "")
-SUPABASE_SITE_URL: str = os.environ.get("SITE_URL", "http://localhost:3000")
+SUPABASE_SITE_URL: str = os.environ.get("SITE_URL") or "http://localhost:3000"
 DATABASE_URL: str = os.environ.get("DATABASE_URL", "")
 
 AUTH_AVAILABLE: bool = bool(SUPABASE_URL and SUPABASE_ANON_KEY)
@@ -23,13 +24,17 @@ def get_supabase():
     global _client
     if _client is None:
         try:
-            from supabase import create_client
+            from supabase import ClientOptions, create_client
         except ImportError:
             raise RuntimeError(
                 "SUPABASE_URL and SUPABASE_ANON_KEY are set but 'supabase' is not installed.\n"
                 "Run: pip install -r requirements-auth.txt"
             )
-        _client = create_client(SUPABASE_URL, SUPABASE_ANON_KEY)
+        _client = create_client(
+            SUPABASE_URL,
+            SUPABASE_ANON_KEY,
+            options=ClientOptions(flow_type="pkce"),
+        )
     return _client
