+++ ourportfolios/pages/home/index.py
@@ -1,24 +1,15 @@
+from typing import Any, cast
+
 import reflex as rx
+
 from ...state.home_state import HomeState
-from ...components.navbar import navbar
+from .layout import home_shell
 from .page_body import page_body
 
 
-@rx.page(route="/home", on_load=HomeState.on_mount)
+@rx.page(route="/home", on_load=cast(Any, HomeState.on_mount))
 def index() -> rx.Component:
-    return rx.box(
-        navbar(),
-        rx.box(
-            page_body(),
-            width="86vw",
-            max_width="90rem",
-            margin="0 auto",
-            padding_y="2rem",
-        ),
-        on_unmount=HomeState.on_unmount,
-        background="#090909",
-        color="white",
-        min_height="100vh",
-        width="100%",
-        overflow_x="hidden",
+    return home_shell(
+        page_body(),
+        on_unmount=cast(Any, HomeState.on_unmount),
     )
