+++ ourportfolios/pages/home/sections/__init__.py
@@ -1,0 +1,5 @@
+"""Home page section composition."""
+
+from .dashboard import page_body
+
+__all__ = ["page_body"]
