+++ ourportfolios/ui/__init__.py
@@ -1,0 +1,34 @@
+"""Shared UI primitives and layout tokens."""
+
+from .layout import clamp_lines, page_frame
+from .tokens import (
+    APP_BG,
+    APP_PANEL_BG,
+    APP_SURFACE_BG,
+    CARD_PREVIEW_HEIGHT,
+    CARD_TEXT_CLAMP_STYLE,
+    HOME_CONTENT_MAX_WIDTH,
+    HOME_CONTENT_WIDTH,
+    HOME_PAGE_HORIZONTAL_PADDING,
+    HOME_PAGE_VERTICAL_PADDING,
+    PAGE_EDGE_PADDING,
+    PAGE_MAX_WIDTH,
+    SECTION_GAP,
+)
+
+__all__ = [
+    "APP_BG",
+    "APP_PANEL_BG",
+    "APP_SURFACE_BG",
+    "CARD_PREVIEW_HEIGHT",
+    "CARD_TEXT_CLAMP_STYLE",
+    "HOME_CONTENT_MAX_WIDTH",
+    "HOME_CONTENT_WIDTH",
+    "HOME_PAGE_HORIZONTAL_PADDING",
+    "HOME_PAGE_VERTICAL_PADDING",
+    "PAGE_EDGE_PADDING",
+    "PAGE_MAX_WIDTH",
+    "SECTION_GAP",
+    "clamp_lines",
+    "page_frame",
+]
