+++ ourportfolios/ui/theme/__init__.py
@@ -1,0 +1,5 @@
+"""Theme package organized by concern for scalable UI styling."""
+
+from .colors import *  # noqa: F401,F403
+from .surfaces import *  # noqa: F401,F403
+from .components import *  # noqa: F401,F403
